\name{dizzysNewInfec-package}
\alias{dizzysNewInfec-package}
\alias{dizzysNewInfec}
\docType{package}
\title{
\packageTitle{dizzysNewInfec}
}
\description{
\packageDescription{dizzysNewInfec}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dizzysNewInfec}
\packageIndices{dizzysNewInfec}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{dizzysNewInfec}

Maintainer: \packageMaintainer{dizzysNewInfec}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
